pyproject.toml
[project]
name = "ai-knowledge-assistant"
version = "0.1.0"
description = "RAG-based AI knowledge assistant with FastAPI, OpenAI, and Pinecone"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "openai>=1.55.0",
    "faiss-cpu>=1.9.0",
    "numpy>=2.0.0",
    "tiktoken>=0.8.0",
    "pypdf>=5.1.0",
    "python-multipart>=0.0.17",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "rank-bm25>=0.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "respx>=0.22.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
